Separate Schema from Conflate
A typical use case is to merge data then validate against a schema.
However, if you want to validate multiple data sets against a single
schema, the API previously forced you to reparse the schema multiple
times, which is suboptimal. Now you can parse the schema once and
store it in an instance of Schema.